Create AuthenticationViewModel, make SignInViewModel and SignUpViewModel extend and use credential manager flow methods
