+++ fundamentals/android/auth-sign-in-with-google/Notes/app/src/main/java/com/notes/app/screens/authentication/CredentialsExt.kt
@@ -1,4 +1,4 @@
-package com.notes.app.screens.sign_up
+package com.notes.app.screens.authentication
 
 import android.util.Patterns
 import java.util.regex.Pattern
