+++ fundamentals/android/auth-sign-in-with-google/Notes/app/src/main/java/com/notes/app/screens/authentication/sign_in/SignInViewModel.kt
@@ -1,10 +1,8 @@
-package com.notes.app.screens.sign_in
+package com.notes.app.screens.authentication.sign_in
 
 import android.util.Log
 import androidx.credentials.Credential
 import androidx.credentials.CustomCredential
-import androidx.credentials.GetCredentialRequest
-import com.google.android.libraries.identity.googleid.GetGoogleIdOption
 import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential
 import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential.Companion.TYPE_GOOGLE_ID_TOKEN_CREDENTIAL
 import com.notes.app.ERROR_TAG
@@ -13,7 +11,7 @@ import com.notes.app.NOTES_LIST_SCREEN
 import com.notes.app.SIGN_UP_SCREEN
 import com.notes.app.UNEXPECTED_CREDENTIAL
 import com.notes.app.model.service.AccountService
-import com.notes.app.screens.NotesAppViewModel
+import com.notes.app.screens.authentication.AuthenticationViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -23,7 +21,7 @@ import javax.inject.Inject
 @HiltViewModel
 class SignInViewModel @Inject constructor(
     private val accountService: AccountService
-) : NotesAppViewModel() {
+) : AuthenticationViewModel() {
     // Backing properties to avoid state updates from other classes
     private val _email = MutableStateFlow("")
     val email: StateFlow<String> = _email.asStateFlow()
@@ -37,17 +35,6 @@ class SignInViewModel @Inject constructor(
 
     fun updatePassword(newPassword: String) {
         _password.value = newPassword
-    }
-
-    fun getCredentialRequest(webClientId: String): GetCredentialRequest {
-        val googleIdOption = GetGoogleIdOption.Builder()
-            .setFilterByAuthorizedAccounts(true)
-            .setServerClientId(webClientId)
-            .build()
-
-        return GetCredentialRequest.Builder()
-            .addCredentialOption(googleIdOption)
-            .build()
     }
 
     fun onSignInClick(openAndPopUp: (String, String) -> Unit) {
